+++ src/main/java/com/wanted/dto/jobposting/JobPostingModificationRequest.java
@@ -1,4 +1,4 @@
-package com.wanted.dto;
+package com.wanted.dto.jobposting;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -15,6 +15,8 @@ public class JobPostingModificationRequest {
     private String title;
     private String imageUrl;
     private String position;
+    private String country;
+    private String region;
     private Long reward;
     private String content;
     private List<String> techStacks;
